backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
Pydantic==2.5.0
ortools==9.8.3296
bcrypt==4.1.2
python-dotenv==1.0.0
email-validator==2.1.0.post1